+++ perimeterx-sdk/src/main/java/com/perimeterx/models/httpmodels/Additional.java
@@ -12,6 +12,7 @@ import com.perimeterx.models.risk.S2SCallReason;
 import com.perimeterx.utils.Constants;
 
 import java.util.Date;
+import java.util.Map;
 import java.util.Objects;
 import java.util.UUID;
 
@@ -85,13 +86,27 @@ public class Additional {
 
     @JsonProperty("request_id")
     public UUID requestId;
+
     @JsonProperty("enforcer_start_time")
     public long enforcerStartTime;
+
     @JsonProperty("risk_start_time")
     public long riskStartTime;
 
     @JsonProperty("cross_tab_session")
     public String pxCtsCookie;
+
+    @JsonProperty("app_user_id")
+    public String appUserId;
+
+    @JsonProperty("jwt_additional_fields")
+    public Map<String, Object> jwtAdditionalFields;
+
+    @JsonProperty("is_sensitive_route")
+    public Boolean isSensitiveRoute;
+
+    @JsonProperty("additional_token_info")
+    public String additionalTokenInfo;
 
     public static Additional fromContext(PXContext ctx) {
         Additional additional = new Additional();
@@ -114,6 +129,10 @@ public class Additional {
         additional.enforcerStartTime = ctx.getEnforcerStartTime();
         additional.riskStartTime = new Date().getTime();
         additional.pxCtsCookie = ctx.getPxCtsCookie();
+        additional.appUserId = ctx.getJwtAppUserId();
+        additional.jwtAdditionalFields = ctx.getJwtAdditionalFields();
+        additional.isSensitiveRoute = ctx.isSensitiveRequest();
+        additional.additionalTokenInfo = ctx.getAdditionalTokenInfo();
 
         setLoginCredentials(ctx, additional);