+++ CHANGELOG.md
@@ -1,4 +1,22 @@
 # Change Log
+## [v6.16.0](https://github.com/PerimeterX/perimeterx-java-sdk/compare/6.16.0...HEAD) (2025-11-12)
+- Added support for data enrichment header feature (`px_data_enrichment_header_name` configuration)
+- Added support for AD user identifiers feature
+- Added `px_secured_pxhd_enabled` configuration option to enable secure flag on `pxhd` cookie
+- Added `is_sensitive_route` to risk api and async activities
+- Added `additional_token_info` to risk api and async activities
+- Updated telemetry activity to new format (`static_config` and `active_config`; `remote_config` is not supported)
+- Updated telemetry activity to include `request_id`
+- Updated captcha page template to newest version
+- Updated dependencies minor and patch versions (major versions unchanged)
+- Changed custom parameters to be of type `Object` instead of `String` to allow more flexibility
+- Changed first party block script in captcha template to end with expected `/captcha.js`
+- Changed `RequestWrapper` to include custom headers in methods that retrieve request headers
+- Fixed possible connection leak issue due to unclosed responses in first party and telemetry requests
+- Fixed first party fuzzing errors by returning 400 on first party requests with URL length > 1000 characters
+
+## [v6.15.1](https://github.com/PerimeterX/perimeterx-java-sdk/compare/6.15.1...HEAD) (2025-09-08)
+- Added additional updateReason RISK to Telemetry flow 
 
 ## [v6.15.0](https://github.com/PerimeterX/perimeterx-java-sdk/compare/6.15.0...HEAD) (2025-09-03)
 - Added Documentation enforcement workflow - verify that the documentation is up to date with the latest changes in the codebase